02-agentic-rag/requirements.txt
langchain>=0.1.0
langgraph>=0.1.0
transformers>=4.37.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24
pydantic>=2.5
rich>=13.7
